Split the rust bindings in two
We now have a metatensor-sys crate which only contains the C API
bindings & code to build/find the C API shared library; and the
main metatensor crate which contains idomatic rust bindings to this
API.

We'll need to release metatensor-sys every time we change metatensor-core,
but the rust bindings can evolve separately
