+++ .github/workflows/rust-tests.yml
@@ -105,8 +105,8 @@ jobs:
         if: matrix.os != 'windows-2019'
         run: |
           ./scripts/package-core.sh
-          cd metatensor
-          cargo publish --dry-run --allow-dirty
+          cd rust/metatensor-sys
+          cargo package --allow-dirty
 
   # check that the C API declarations are correctly used by Rust and Python
   prevent-bitrot:
