+++ metatensor-core/Cargo.toml
@@ -9,6 +9,7 @@ exclude = [
 ]
 
 [lib]
+crate-type = ["cdylib", "staticlib"]
 name = "metatensor"
 bench = false
