+++ metatensor-core/tests/data.npz
