+++ metatensor-core/tests/keys.npy
