+++ metatensor-torch/tests/CMakeLists.txt
@@ -38,7 +38,7 @@ foreach(_file_ ${ALL_TESTS})
     get_filename_component(_name_ ${_file_} NAME_WE)
     add_executable(torch-${_name_} ${_file_})
     target_link_libraries(torch-${_name_} metatensor_torch catch)
-    target_compile_definitions(torch-${_name_} PRIVATE "-DDATA_NPZ=\"${CMAKE_CURRENT_SOURCE_DIR}/../../metatensor/tests/data.npz\"")
+    target_compile_definitions(torch-${_name_} PRIVATE "-DDATA_NPZ=\"${CMAKE_CURRENT_SOURCE_DIR}/../../metatensor-core/tests/data.npz\"")
 
     add_test(
         NAME torch-${_name_}
