+++ rust/metatensor-sys/.gitignore
