+++ rust/metatensor-sys/build.rs
@@ -2,7 +2,7 @@ use std::path::PathBuf;
 use std::process::Command;
 
 fn main() {
-    let mut metatensor_core = PathBuf::from("../metatensor-core");
+    let mut metatensor_core = PathBuf::from("../../metatensor-core");
 
     // setting DESTDIR when building with make will cause the install to be in a
     // different directory than the expected one ($OUT_DIR/lib)
@@ -11,24 +11,19 @@ fn main() {
     let mut cargo_toml = metatensor_core.clone();
     cargo_toml.push("Cargo.toml");
 
-    // when packaging for crates.io, the metatensor-core symlink is not included.
-    // instead, we manually package metatensor-core as a .crate files (actually a
-    // .tar.gz file), and unpack it here. We then use cmake to build the code as
-    // if it was a standard C library (and cmake calls back cargo to build the
-    // rust code in metatensor-core)
+    // when packaging for crates.io, the metatensor-core symlink is not
+    // included. instead, we manually package metatensor-core as a .tar.gz file
+    // and unpack it here. We then use cmake to build the code as if it was a
+    // standard C library (and cmake calls back cargo to build the rust code in
+    // metatensor-core)
     if !cargo_toml.is_file() {
         let cmake_exe = which::which("cmake").expect("could not find cmake");
 
-        let all_crate_files = glob::glob("metatensor-core-cxx-*.tar.gz")
-            .expect("bad pattern")
-            .flatten()
-            .collect::<Vec<_>>();
-
-        if all_crate_files.len() != 1 {
-            panic!("could not find the metatensor-core-cxx tarball, run script/update-core.sh");
+        let mut core_archive = std::env::current_dir().expect("missing cwd");
+        core_archive.push(format!("metatensor-core-cxx-{}.tar.gz", env!("CARGO_PKG_VERSION")));
+        if !core_archive.exists() {
+             panic!("missing metatensor-core-cxx archive, run scripts/package-core.sh");
         }
-        let mut crate_file = std::env::current_dir().expect("missing cwd");
-        crate_file.push(&all_crate_files[0]);
 
         metatensor_core = PathBuf::from(std::env::var("OUT_DIR").expect("missing OUT_DIR"));
 
@@ -36,14 +31,16 @@ fn main() {
             .arg("-E")
             .arg("tar")
             .arg("xf")
-            .arg(&crate_file)
+            .arg(&core_archive)
             .current_dir(&metatensor_core)
             .status()
             .expect("failed to unpack metatensor-core");
 
-        let crate_dir = crate_file.file_name().expect("file name").to_str().expect("UTF8 error");
-        let splitted = crate_dir.split('.').collect::<Vec<_>>();
+        let core_dir = core_archive.file_name().expect("file name").to_str().expect("UTF8 error");
+        let splitted = core_dir.split('.').collect::<Vec<_>>();
         metatensor_core.push(splitted[..splitted.len() - 2].join("."));
+    } else {
+        println!("cargo:rerun-if-changed={}", metatensor_core.display());
     }
 
     let install_dir = cmake::Config::new(&metatensor_core)
@@ -69,5 +66,4 @@ fn main() {
     }
 
     println!("cargo:rerun-if-changed=build.rs");
-    println!("cargo:rerun-if-changed={}", metatensor_core.display());
 }
