+++ rust/metatensor/.gitignore
@@ -1,0 +1,1 @@
+metatensor-core-cxx-*.tar.gz
