+++ rust/metatensor/Cargo.toml
@@ -3,14 +3,9 @@ name = "metatensor"
 version = "0.1.3"
 edition = "2021"
 rust-version = "1.65"
-include = [
-    "build.rs",
-    "src/",
-    "metatensor-core-cxx-*.tar.gz",
-]
 
 description = "Self-describing sparse tensor data format for atomistic machine learning and beyond"
-readme = "../README.md"
+readme = "../../README.md"
 homepage = "https://lab-cosmo.github.io/metatensor/latest/"
 repository = "https://github.com/lab-cosmo/metatensor"
 license = "BSD-3-Clause"
@@ -19,21 +14,15 @@ license = "BSD-3-Clause"
 bench = false
 
 [dependencies]
+metatensor-sys = {version = "0.1.3", path="../metatensor-sys"}
+
 once_cell = "1"
 smallvec = {version = "1", features = ["union"]}
 ndarray = {version = "0.15"}
 rayon = {version = "1", optional = true}
 
+
 [features]
 default = []
 # use the static build of metatensor-core instead of the shared one
-static = []
-
-[build-dependencies]
-# we want a recent version of the cmake crate
-cmake = "0.1.49"
-which = "5"
-glob = "0.3"
-
-# This is the last version supporting rustc 1.65
-home = "=0.5.5"
+static = ["metatensor-sys/static"]
