+++ rust/metatensor/src/block/block_ref.rs
