+++ rust/metatensor/src/block/mod.rs
