+++ rust/metatensor/src/data/array.rs
