+++ rust/metatensor/src/io/labels.rs
