+++ rust/metatensor/src/io/mod.rs
