+++ rust/metatensor/src/io/tensor.rs
