+++ rust/metatensor/tests/components-to-properties.rs
