+++ rust/metatensor/tests/keys-to-properties.rs
