+++ rust/metatensor/tests/keys-to-samples.rs
